demos/ecb_hr_expert_agent/pyproject.toml
[project]
name = "ecb-hr-expert-agent"
version = "0.0.0"
requires-python = ">=3.13"
dependencies = [
    "dead-parrot",
    "dspy>=3.1.3",
    "dspy-temporal>=0.0.6",
    "fastapi>=0.135.3",
    "fastmcp>=3.2.3",
    "python-dotenv>=1.2.2",
    "temporalio>=1.25.0",
    "uvicorn>=0.44.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
dead-parrot = { workspace = true }
